restructure Timeout Exception
* test passed
